pyproject.toml
[project]
name = "hikcam-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "albumentations>=1.4.18",
    "paddleocr==2.9",
    "paho-mqtt>=2.1.0",
    "pip>=25.0.1",
    "setuptools>=75.3.2",
]
